Refactor services directory to be per-file
