+++ apps/back-end/src/server/routes/blogs/getBlogById.ts
@@ -1,7 +1,7 @@
 import type { Router } from "express";
 
 import { getConnection } from "src/database/connection";
-import { selectBlogView } from "src/services/blogs";
+import loadBlogView from "src/services/blogs/loadBlogView";
 import resourceNotFoundError from "src/utility/errors/resourceNotFoundError";
 import handleEndpointMiddleware from "src/utility/handlers/handleEndpointMiddleware";
 import validateUUID from "src/utility/handlers/validateUUID";
@@ -11,7 +11,7 @@ function getBlogById(blogs: Router) {
     "/:blogId",
     handleEndpointMiddleware<{ blogId: string }>(async (request, response) => {
       const connection = getConnection();
-      const blog = await selectBlogView(connection, request.params.blogId);
+      const blog = await loadBlogView(connection, request.params.blogId);
 
       if (blog === null) {
         throw resourceNotFoundError("blog", request.params.blogId);
