+++ apps/back-end/src/services/blogs/helpers/BlogEndpointIds.ts
@@ -1,0 +1,4 @@
+export interface BlogEndpointIds {
+  blogId: string;
+  editorId: string;
+}
