+++ apps/back-end/src/services/userSessions/createUserSession.ts
@@ -3,12 +3,10 @@ import type { CreateUserSessionData, UserSession } from "@lexicon/models";
 import type { Connection } from "src/database/connection";
 
 import { addDaysToDate } from "@alextheman/utility";
-import { parseUserSession } from "@lexicon/models";
 
 import insertUserSession from "src/models/userSessions/insertUserSession";
-import updateUserSession from "src/models/userSessions/updateUserSession";
 
-export async function createUserSession(
+async function createUserSession(
   connection: Connection,
   data: CreateUserSessionData,
 ): Promise<UserSession> {
@@ -18,13 +16,8 @@ export async function createUserSession(
     createdAt: today,
     expiresAt: data.expiresAt ?? addDaysToDate(today, 7),
   });
-  return parseUserSession(userSession);
+
+  return userSession;
 }
 
-export async function expireUserSession(
-  connection: Connection,
-  sessionId: string,
-): Promise<UserSession | null> {
-  const userSession = await updateUserSession(connection, sessionId, { expiresAt: new Date() });
-  return parseUserSession(userSession);
-}
+export default createUserSession;
