+++ apps/back-end/src/services/users/editUserProfile.ts
@@ -1,0 +1,24 @@
+import type { User, UserProfileData } from "@lexicon/models";
+
+import type { Connection } from "src/database/connection";
+
+import { az } from "@alextheman/utility";
+import z from "zod";
+
+import updateUser from "src/models/users/updateUser";
+
+async function editUserProfile(
+  connection: Connection,
+  userId: string,
+  data: UserProfileData,
+): Promise<User | null> {
+  const user = await updateUser(connection, userId, data);
+
+  if (user !== null) {
+    return { ...user, dateOfBirth: az.with(z.coerce.date().nullable()).parse(user.dateOfBirth) };
+  }
+
+  return null;
+}
+
+export default editUserProfile;
